+++ src/home.c
@@ -14,20 +14,20 @@
 //#define START
 
 /* Function prototypes */
-static void	generate_farts(struct game *cur_game, struct worldmap *main_map);
-static void	copy_fart(struct worldmap *main_map, struct worldmap *fart, int row, int col);
-static void	setup_tiles(struct game *cur_game);
-static void	draw_title_tiles(struct game *cur_game);
-static void	draw_title_tile(struct game *cur_game, int i, int j);
-static void	draw_title_screen(struct game *cur_game, int status);
-static void	draw_bg(struct game *cur_game);
-static void	save_screen(struct game *cur_game, struct worldmap *map, struct player *cur_player);
-static SDL_bool	load_screen(struct game *cur_game, struct worldmap *map, struct player *cur_player);
-static void	opts_screen(struct game *cur_game);
-static void	draw_saveload_screen(struct game *cur_game, SDL_bool *saves, SDL_bool save);
-static void	draw_opts_screen(struct game *cur_game);
-static void	draw_yesno_screen(struct game *cur_game, char *message);
-
+static void		generate_farts(struct game *cur_game, struct worldmap *main_map);
+static void		copy_fart(struct worldmap *main_map, struct worldmap *fart, int row, int col);
+static void		setup_tiles(struct game *cur_game);
+static void		draw_title_tiles(struct game *cur_game);
+static void		draw_title_tile(struct game *cur_game, int i, int j);
+static void		draw_title_screen(struct game *cur_game, int status);
+static void		draw_bg(struct game *cur_game);
+static void		save_screen(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+static SDL_bool		load_screen(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+static void		opts_screen(struct game *cur_game);
+static void		draw_saveload_screen(struct game *cur_game, SDL_bool *saves, SDL_bool save);
+static void		draw_opts_screen(struct game *cur_game);
+static void		draw_yesno_screen(struct game *cur_game, char *message);
+static SDL_bool		change_resolution(struct game *cur_game, int w, int h);
 
 void
 title_screen(struct game *cur_game, struct worldmap *map, struct player *cur_player, int status)
@@ -86,9 +86,9 @@ title_screen(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 			new_game = QUITTING_GAME;
 			finished = SDL_TRUE;
 		} else if (event.type == SDL_MOUSEBUTTONDOWN && event.button.button == SDL_BUTTON_LEFT) {
-			x = event.button.x; y = event.button.y;
-			if (x >= 123  * cur_game->screen.scale_x && x <= 219  * cur_game->screen.scale_x &&
-			    y >= 153 * cur_game->screen.scale_y && y <= 162 * cur_game->screen.scale_y) {
+			x = event.button.x / cur_game->screen.scale_x;
+			y = event.button.y / cur_game->screen.scale_y;
+			if (x >= 123 && x <= 219 && y >= 153 && y <= 162) {
 				/* new game or save game */
 				if (status == STARTING_GAME) {
 					new_game = STARTING_GAME;
@@ -98,8 +98,7 @@ title_screen(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 					finished = SDL_TRUE;
 					save_screen(cur_game, map, cur_player);
 				}
-			} else if (x >= 123 * cur_game->screen.scale_x && x <= 219 * cur_game->screen.scale_x &&
-				   y >= 175 * cur_game->screen.scale_y && y <= 183 * cur_game->screen.scale_y) {
+			} else if (x >= 123 && x <= 219 && y >= 175 && y <= 183) {
 				/* continue */
 				if (status == STARTING_GAME) {
 					new_game = CONTINUE_GAME;
@@ -108,13 +107,11 @@ title_screen(struct game *cur_game, struct worldmap *map, struct player *cur_pla
 					new_game = GAME_IN_PROGRESS;
 					finished = SDL_TRUE;
 				}
-			} else if (x >= 123 * cur_game->screen.scale_x && x <= 195 * cur_game->screen.scale_x &&
-				   y >= 193 * cur_game->screen.scale_y && y <= 203 * cur_game->screen.scale_y) {
+			} else if (x >= 123 && x <= 195 && y >= 193 && y <= 203) {
 				/* options */
 				opts_screen(cur_game);
 				continue;
-			} else if (x >= 123 * cur_game->screen.scale_x && x <= 161 * cur_game->screen.scale_x &&
-				   y >= 213 * cur_game->screen.scale_y && y <= 222 * cur_game->screen.scale_y) {
+			} else if (x >= 123 && x <= 161 && y >= 213 && y <= 222) {
 				/* exit */
 				finished = SDL_TRUE;
 				new_game = QUITTING_GAME;
@@ -215,7 +212,8 @@ setup_tiles(struct game *cur_game)
 	/* Create texture if necessary */
 	if (title_screen_tex == NULL) {
 		title_screen_tex = SDL_CreateTexture(cur_game->screen.renderer, SDL_PIXELFORMAT_RGBA8888,
-						 SDL_TEXTUREACCESS_TARGET, cur_game->screen.w, cur_game->screen.h);
+						 SDL_TEXTUREACCESS_TARGET, 1280, 720);
+		SDL_SetTextureBlendMode(title_screen_tex, SDL_BLENDMODE_BLEND);
 	}
 	/* Output all tiles to texture */
 	draw_title_tiles(cur_game);
@@ -230,15 +228,14 @@ draw_title_tiles(struct game *cur_game)
 	
 	/* Render to texture */
 	SDL_SetRenderTarget(cur_game->screen.renderer, title_screen_tex);
-	SDL_SetTextureBlendMode(title_screen_tex, SDL_BLENDMODE_BLEND);
-	draw_rect(cur_game, 0, 0, cur_game->screen.w, cur_game->screen.h, SDL_TRUE, black, SDL_FALSE, NULL);
+	draw_rect(cur_game, 0, 0, 1280, 720, SDL_TRUE, black, SDL_FALSE, NULL);
 	for (i = 0; i < 80; i++) {
 		for (j = 0; j < 45; j++) {
-			draw_tile(cur_game, i * 16 * cur_game->screen.scale_x, j * 16 * cur_game->screen.scale_y,
-				  16 * cur_game->screen.scale_x, 16 * cur_game->screen.scale_y, tile[i][j], 128);
+			draw_tile(cur_game, i * 16, j * 16,
+				  16, 16, tile[i][j], 128);
 		}
 	}
-	SDL_SetRenderTarget(cur_game->screen.renderer, NULL);
+	SDL_SetRenderTarget(cur_game->screen.renderer, cur_game->screen.output);
 }
 
 static void
@@ -248,10 +245,9 @@ draw_title_tile(struct game *cur_game, int i, int j)
 	
 	/* Render to texture */
 	SDL_SetRenderTarget(cur_game->screen.renderer, title_screen_tex);
-	SDL_SetTextureBlendMode(title_screen_tex, SDL_BLENDMODE_BLEND);
-	draw_rect(cur_game, i * 16 * cur_game->screen.scale_x, j * 16 * cur_game->screen.scale_y, 16 * cur_game->screen.scale_x, 16 * cur_game->screen.scale_y, SDL_TRUE, black, SDL_FALSE, NULL);
-	draw_tile(cur_game, i * 16 * cur_game->screen.scale_x, j * 16 * cur_game->screen.scale_y, 16 * cur_game->screen.scale_x, 16 * cur_game->screen.scale_y, tile[i][j], 128);
-	SDL_SetRenderTarget(cur_game->screen.renderer, NULL);
+	draw_rect(cur_game, i * 16, j * 16, 16, 16, SDL_TRUE, black, SDL_FALSE, NULL);
+	draw_tile(cur_game, i * 16, j * 16, 16, 16, tile[i][j], 128);
+	SDL_SetRenderTarget(cur_game->screen.renderer, cur_game->screen.output);
 }
 
 static void
@@ -260,28 +256,28 @@ draw_title_screen(struct game *cur_game, int status)
 	/* Draw background */
 	draw_bg(cur_game);	
 	/* draw game title */
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 100 * cur_game->screen.scale_y, "MAPGAME");
+	draw_sentence(cur_game, 100, 100, "MAPGAME");
 	/* draw new game */
 	if (status == STARTING_GAME) {
-		draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "NEW GAME");
+		draw_small_sentence(cur_game, 120, 150, "NEW GAME");
 	} else {
-		draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "SAVE GAME");
+		draw_small_sentence(cur_game, 120, 150, "SAVE GAME");
 	}
 	/* draw continue */
 	if (status == STARTING_GAME) {
-		draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 170 * cur_game->screen.scale_y, "LOAD GAME");
+		draw_small_sentence(cur_game, 120, 170, "LOAD GAME");
 	} else {
-		draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 170 * cur_game->screen.scale_y, "CONTINUE");
+		draw_small_sentence(cur_game, 120, 170, "CONTINUE");
 	}
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 190 * cur_game->screen.scale_y, "OPTIONS");
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 210 * cur_game->screen.scale_y, "EXIT");
+	draw_small_sentence(cur_game, 120, 190, "OPTIONS");
+	draw_small_sentence(cur_game, 120, 210, "EXIT");
 }
 
 static void
 draw_bg(struct game *cur_game)
 {
 	int i, j;
-	SDL_Rect rect = { 0, 0, cur_game->screen.w, cur_game->screen.h };
+	SDL_Rect rect = { 0, 0, 1280, 720 };
 	
 	/* Have a random chance to change a tile */
 	if (rand_num(0, 99) >= 49) {
@@ -337,15 +333,16 @@ save_screen(struct game *cur_game, struct worldmap *map, struct player *cur_play
 					break;
 			}
 		} else if (event.type == SDL_MOUSEBUTTONDOWN && event.button.button == SDL_BUTTON_LEFT) {
-			x = event.button.x; y = event.button.y;
-			if (x >= 122 * cur_game->screen.scale_x && x <= 343 * cur_game->screen.scale_x) {
-				if (y >= 152 * cur_game->screen.scale_y && y <= 163 * cur_game->screen.scale_y) {
+			x = event.button.x / cur_game->screen.scale_x;
+			y = event.button.y / cur_game->screen.scale_y;
+			if (x >= 122 && x <= 343) {
+				if (y >= 152 && y <= 163) {
 					finished = save_all(cur_game, map, cur_player, 0);
-				} else if (y >= 172 * cur_game->screen.scale_y && y <= 183 * cur_game->screen.scale_y) {
+				} else if (y >= 172 && y <= 183) {
 					finished = save_all(cur_game, map, cur_player, 1);
-				} else if (y >= 192 * cur_game->screen.scale_y && y <= 204 * cur_game->screen.scale_y) {
+				} else if (y >= 192 && y <= 204) {
 					finished = save_all(cur_game, map, cur_player, 2);
-				} else if (y >= 212 * cur_game->screen.scale_y && y <= 224 * cur_game->screen.scale_y) {
+				} else if (y >= 212 && y <= 224) {
 					finished = save_all(cur_game, map, cur_player, 3);
 				}
 			}
@@ -411,21 +408,22 @@ load_screen(struct game *cur_game, struct worldmap *map, struct player *cur_play
 					break;
 			}
 		} else if (event.type == SDL_MOUSEBUTTONDOWN && event.button.button == SDL_BUTTON_LEFT) {
-			x = event.button.x; y = event.button.y;
-			if (x >= 122 * cur_game->screen.scale_x && x <= 343 * cur_game->screen.scale_x) {
-				if (y >= 152 * cur_game->screen.scale_y && y <= 163 * cur_game->screen.scale_y) {
+			x = event.button.x / cur_game->screen.scale_x;
+			y = event.button.y / cur_game->screen.scale_y;
+			if (x >= 122 && x <= 343) {
+				if (y >= 152 && y <= 163 && saves[0] == SDL_TRUE) {
 					load_all(cur_game, map, cur_player, 0);
 					loaded = SDL_TRUE;
 					finished = SDL_TRUE;
-				} else if (y >= 172 * cur_game->screen.scale_x && y <= 183 * cur_game->screen.scale_y) {
+				} else if (y >= 172 && y <= 183 && saves[1] == SDL_TRUE) {
 					load_all(cur_game, map, cur_player, 1);
 					loaded = SDL_TRUE;
 					finished = SDL_TRUE;
-				} else if (y >= 192 * cur_game->screen.scale_x && y <= 204 * cur_game->screen.scale_y) {
+				} else if (y >= 192 && y <= 204 && saves[2] == SDL_TRUE) {
 					load_all(cur_game, map, cur_player, 2);
 					loaded = SDL_TRUE;
 					finished = SDL_TRUE;
-				} else if (y >= 212 * cur_game->screen.scale_x && y <= 224 * cur_game->screen.scale_y) {
+				} else if (y >= 212 && y <= 224 && saves[3] == SDL_TRUE) {
 					load_all(cur_game, map, cur_player, 3);
 					loaded = SDL_TRUE;
 					finished = SDL_TRUE;
@@ -464,83 +462,99 @@ opts_screen(struct game *cur_game)
 					break;
 			}
 		} else if (event.type == SDL_MOUSEBUTTONDOWN && event.button.button == SDL_BUTTON_LEFT) {
-			x = event.button.x; y = event.button.y;
-			if (x >= 100 * cur_game->screen.scale_x && x <= 165 * cur_game->screen.scale_x &&
-			    y >= 574 * cur_game->screen.scale_y && y <= 586 * cur_game->screen.scale_y) {
-			    	return;
-			} else if (x >= 190 * cur_game->screen.scale_x && x <= 346 * cur_game->screen.scale_x) {
-				if (y >= 185 * cur_game->screen.scale_y && y <= 199 * cur_game->screen.scale_y &&
-				    cur_game->screen.w != 960 && cur_game->screen.h != 540) {
-					/* 960x540 */
-					cur_game->screen.w = 960;
-					cur_game->screen.h = 540;
-					cur_game->screen.scale_x = 960.0/1280.0;
-					cur_game->screen.scale_y = 960.0/1280.0;
-					changed = SDL_TRUE;
-				} else if (y >= 214 * cur_game->screen.scale_y && y <= 227 * cur_game->screen.scale_y &&
-					   cur_game->screen.w != 1280 && cur_game->screen.h != 720) {
+			x = event.button.x / cur_game->screen.scale_x;
+			y = event.button.y / cur_game->screen.scale_y;
+			if (x >= 101 && x <= 275) {
+				/* Clicked on main display options */
+				if (y >= 180 && y <= 195) {
+					/* fullscreen */
+					if (cur_game->screen.displaymode == 2) {
+						cur_game->screen.displaymode = 0;
+						changed = SDL_TRUE;
+					}
+				} else if (y >= 200 && y <= 215) {
+					/* window */
+					if (cur_game->screen.displaymode != 2) {
+						cur_game->screen.displaymode = 2;
+						changed = SDL_TRUE;
+					}
+				} else if (y >= 240 && y <= 255) {
+					/* vsync on */
+					if (cur_game->screen.vsync == SDL_FALSE) {
+						cur_game->screen.vsync = SDL_TRUE;
+						changed = SDL_TRUE;
+					}
+				} else if (y >= 260 && y <= 275) {
+					/* vsync off */
+					if (cur_game->screen.vsync == SDL_TRUE) {
+						cur_game->screen.vsync = SDL_FALSE;
+						changed = SDL_TRUE;
+					}
+				} else if (y >= 300 && y <= 315) {
+					/* scanlines on */
+					if (cur_game->screen.scanlines_on == SDL_FALSE) {
+						cur_game->screen.scanlines_on = SDL_TRUE;
+					}
+				} else if (y >= 320 && y <= 335) {
+					/* scanlines off */
+					if (cur_game->screen.scanlines_on == SDL_TRUE) {
+						cur_game->screen.scanlines_on = SDL_FALSE;
+					}
+				}	
+			} else if (x >= 420 && x <= 545) {
+				/* Clicked on resolutions */
+				if (y >= 180 && y <= 195) {
+					/* Default */
+					if (cur_game->screen.displaymode == 1) {
+						cur_game->screen.displaymode = 0;
+						changed = SDL_TRUE;
+					} else if (cur_game->screen.displaymode == 2) {
+						changed = change_resolution(cur_game, 1280, 720);
+					}
+				} else if (y >= 200 && y <= 215) {
 					/* 1280x720 */
-					cur_game->screen.w = 1280;
-					cur_game->screen.h = 720;
-					cur_game->screen.scale_x = 1.0;
-					cur_game->screen.scale_y = 1.0;
-					changed = SDL_TRUE;
-				} else if (y >= 243 * cur_game->screen.scale_y && y <= 258 * cur_game->screen.scale_y &&
-					   cur_game->screen.w != 1600 && cur_game->screen.h != 900) {
+					changed = change_resolution(cur_game, 1280, 720);
+				} else if (y >= 220 && y <= 235) {
+					/* 1360x768 */
+					changed = change_resolution(cur_game, 1360, 768);
+				} else if (y >= 240 && y <= 255) {
+					/* 1366x768 */
+					changed = change_resolution(cur_game, 1366, 768);
+				} else if (y >= 260 && y <= 275) {
+					/* 1440x900 */
+					changed = change_resolution(cur_game, 1440, 900);
+				} else if (y >= 280 && y <= 295) {
 					/* 1600x900 */
-					cur_game->screen.w = 1600;
-					cur_game->screen.h = 900;
-					cur_game->screen.scale_x = 1600.0/1280.0;
-					cur_game->screen.scale_y = 1600.0/1280.0;
-					changed = SDL_TRUE;
-				} else if (y >= 275 * cur_game->screen.scale_y && y <= 286 * cur_game->screen.scale_y &&
-					   cur_game->screen.w != 1920 && cur_game->screen.h != 1080) {
+					changed = change_resolution(cur_game, 1600, 900);
+				} else if (y >= 300 && y <= 315) {
+					/* 1680x1050 */
+					changed = change_resolution(cur_game, 1680, 1050);
+				} else if (y >= 320 && y <= 335) {
 					/* 1920x1080 */
-					cur_game->screen.w = 1920;
-					cur_game->screen.h = 1080;
-					cur_game->screen.scale_x = 1920.0/1280.0;
-					cur_game->screen.scale_y = 1920.0/1280.0;
-					changed = SDL_TRUE;
-				} else if (y >= 334 * cur_game->screen.scale_y && y <= 348 * cur_game->screen.scale_y &&
-					   cur_game->screen.fullscreen == SDL_FALSE) {
-					/* Window */
-					cur_game->screen.fullscreen = SDL_TRUE;
-					changed = SDL_TRUE;
-				} else if (y >= 364 * cur_game->screen.scale_y && y <= 378 * cur_game->screen.scale_y &&
-					   cur_game->screen.fullscreen == SDL_TRUE) {
-					cur_game->screen.fullscreen = SDL_FALSE;
-					changed = SDL_TRUE;
-				} else if (y >= 423 * cur_game->screen.scale_y && y <= 436 * cur_game->screen.scale_y &&
-					   cur_game->screen.vsync == SDL_FALSE) {
-					/* Vsync on */
-					cur_game->screen.vsync = SDL_TRUE;
-					changed = SDL_TRUE;
-				} else if (y >= 453 * cur_game->screen.scale_y && y <= 467 * cur_game->screen.scale_y &&
-					   cur_game->screen.vsync == SDL_TRUE) {
-					/* Vsync off */
-					cur_game->screen.vsync = SDL_FALSE;
-					changed = SDL_TRUE;
-				} else if (y >= 514 * cur_game->screen.scale_y && y <= 527 * cur_game->screen.scale_y &&
-					   cur_game->screen.scanlines_on == SDL_FALSE) {
-					/* Scanlines on */
-					cur_game->screen.scanlines_on = SDL_TRUE;
-					changed = SDL_TRUE;
-				} else if (y >= 543 * cur_game->screen.scale_y && y <= 557 * cur_game->screen.scale_y &&
-					   cur_game->screen.scanlines_on == SDL_TRUE) {
-					/* Scanlines off */
-					cur_game->screen.scanlines_on = SDL_FALSE;
-					changed = SDL_TRUE;
-				}
+					changed = change_resolution(cur_game, 1920, 1080);
+				} else if (y >= 340 && y <= 355) {
+					/* 1920x1200 */
+					changed = change_resolution(cur_game, 1920, 1200);
+				} else if (y >= 360 && y <= 375) {
+					/* 2560x1440 */
+					changed = change_resolution(cur_game, 2560, 1440);
+				} else if (y >= 380 && y <= 395) {
+					/* 2560x1600 */
+					changed = change_resolution(cur_game, 2560, 1600);
+				} else if (y >= 400 && y <= 415) {
+					/* 3840x2160 */
+					changed = change_resolution(cur_game, 3840, 2160);
+				} 
 			}
 		}
 		if (changed == SDL_TRUE) {
+			/* Reset up title screen texture */
+			SDL_DestroyTexture(title_screen_tex);
+			title_screen_tex = NULL;
 			/* change video settings - reinitialize display */
 			display_quit(cur_game);
 			display_init(cur_game);
 			changed = SDL_FALSE;
-			/* Reset up title screen texture */
-			SDL_DestroyTexture(title_screen_tex);
-			title_screen_tex = NULL;
 			setup_tiles(cur_game);
 			save_opts(cur_game);
 		}
@@ -554,19 +568,19 @@ draw_saveload_screen(struct game *cur_game, SDL_bool *saves, SDL_bool save)
 	draw_bg(cur_game);	
 	/* draw save or load header */
 	if (save == SDL_TRUE) {
-		draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 100 * cur_game->screen.scale_y, "SAVE GAME");
+		draw_sentence(cur_game, 100, 100, "SAVE GAME");
 	} else {
-		draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 100 * cur_game->screen.scale_y, "LOAD GAME");
+		draw_sentence(cur_game, 100, 100, "LOAD GAME");
 	}
 	/* draw saves */
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "GAME 1");
-	if (saves[0]) draw_small_sentence(cur_game, 220 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "SAVE EXISTS");
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 170 * cur_game->screen.scale_y, "GAME 2");
-	if (saves[1]) draw_small_sentence(cur_game, 220 * cur_game->screen.scale_x, 170 * cur_game->screen.scale_y, "SAVE EXISTS");
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 190 * cur_game->screen.scale_y, "GAME 3");
-	if (saves[2]) draw_small_sentence(cur_game, 220 * cur_game->screen.scale_x, 190 * cur_game->screen.scale_y, "SAVE EXISTS");
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 210 * cur_game->screen.scale_y, "GAME 4");
-	if (saves[3]) draw_small_sentence(cur_game, 220 * cur_game->screen.scale_x, 210 * cur_game->screen.scale_y, "SAVE EXISTS");
+	draw_small_sentence(cur_game, 120, 150, "GAME 1");
+	if (saves[0]) draw_small_sentence(cur_game, 220, 150, "SAVE EXISTS");
+	draw_small_sentence(cur_game, 120, 170, "GAME 2");
+	if (saves[1]) draw_small_sentence(cur_game, 220, 170, "SAVE EXISTS");
+	draw_small_sentence(cur_game, 120, 190, "GAME 3");
+	if (saves[2]) draw_small_sentence(cur_game, 220, 190, "SAVE EXISTS");
+	draw_small_sentence(cur_game, 120, 210, "GAME 4");
+	if (saves[3]) draw_small_sentence(cur_game, 220, 210, "SAVE EXISTS");
 }
 
 SDL_bool
@@ -602,13 +616,12 @@ yesno_screen(struct game *cur_game, char *message)
 					break;
 			}
 		} else if (event.type == SDL_MOUSEBUTTONDOWN && event.button.button == SDL_BUTTON_LEFT) {
-			x = event.button.x; y = event.button.y;
-			if (x >= 121 * cur_game->screen.scale_x && x <= 154 * cur_game->screen.scale_x &&
-			    y >= 152 * cur_game->screen.scale_y && y <= 163 * cur_game->screen.scale_y) {
+			x = event.button.x / cur_game->screen.scale_x;
+			y = event.button.y / cur_game->screen.scale_y;
+			if (x >= 121 && x <= 154 && y >= 152 && y <= 163) {
 				yes_or_no = SDL_TRUE;
 				finished = SDL_TRUE;
-			} else if (x >= 121 * cur_game->screen.scale_x && x <= 141 * cur_game->screen.scale_x &&
-				   y >= 173 * cur_game->screen.scale_y && y <= 182 * cur_game->screen.scale_y) {
+			} else if (x >= 121 && x <= 141 && y >= 173 && y <= 182) {
 				yes_or_no = SDL_FALSE;
 				finished = SDL_TRUE;
 			}
@@ -623,34 +636,50 @@ draw_opts_screen(struct game *cur_game)
 	/* Draw background */
 	draw_bg(cur_game);	
 	/* draw options header */
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 100 * cur_game->screen.scale_y, "GAME OPTIONS");
+	draw_sentence(cur_game, 100, 100, "DISPLAY OPTIONS");
 	
-	/* draw saves */
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "RESOLUTION");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 180 * cur_game->screen.scale_y, "    960x540");
-	if (cur_game->screen.w == 960 && cur_game->screen.h == 540) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 180 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 210 * cur_game->screen.scale_y, "    1280x720 (RECOMMENDED)");
-	if (cur_game->screen.w == 1280 && cur_game->screen.h == 720) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 210 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 240 * cur_game->screen.scale_y, "    1600x900");
-	if (cur_game->screen.w == 1600 && cur_game->screen.h == 900) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 240 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 270 * cur_game->screen.scale_y, "    1920x1080");
-	if (cur_game->screen.w == 1920 && cur_game->screen.h == 1080) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 270 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 300 * cur_game->screen.scale_y, "MODE");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 330 * cur_game->screen.scale_y, "    FULLSCREEN");
-	if (cur_game->screen.fullscreen == SDL_TRUE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 330 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 360 * cur_game->screen.scale_y, "    WINDOWED");
-	if (cur_game->screen.fullscreen == SDL_FALSE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 360 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 390 * cur_game->screen.scale_y, "VSYNC");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 420 * cur_game->screen.scale_y, "    ON");
-	if (cur_game->screen.vsync == SDL_TRUE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 420 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 450 * cur_game->screen.scale_y, "    OFF");
-	if (cur_game->screen.vsync == SDL_FALSE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 450 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 480 * cur_game->screen.scale_y, "SCANLINES");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 510 * cur_game->screen.scale_y, "    ON (RECOMMENDED)");
-	if (cur_game->screen.scanlines_on == SDL_TRUE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 510 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 540 * cur_game->screen.scale_y, "    OFF");
-	if (cur_game->screen.scanlines_on == SDL_FALSE) draw_sentence(cur_game, 120 * cur_game->screen.scale_x, 540 * cur_game->screen.scale_y, "  >");
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 570 * cur_game->screen.scale_y, "DONE");
+	/* draw display options */
+	draw_small_sentence(cur_game, 100, 160, "DISPLAY MODE");
+	draw_small_sentence(cur_game, 120, 180, "    FULLSCREEN");
+	if (cur_game->screen.displaymode == 0 || cur_game->screen.displaymode == 1) draw_small_sentence(cur_game, 120, 180, "  >");
+	draw_small_sentence(cur_game, 120, 200, "    WINDOW");
+	if (cur_game->screen.displaymode == 2) draw_small_sentence(cur_game, 120, 200, "  >");
+	draw_small_sentence(cur_game, 100, 220, "VSYNC");
+	draw_small_sentence(cur_game, 120, 240, "    ON");
+	if (cur_game->screen.vsync == SDL_TRUE) draw_small_sentence(cur_game, 120, 240, "  >");
+	draw_small_sentence(cur_game, 120, 260, "    OFF");
+	if (cur_game->screen.vsync == SDL_FALSE) draw_small_sentence(cur_game, 120, 260, "  >");
+	draw_small_sentence(cur_game, 100, 280, "SCANLINES");
+	draw_small_sentence(cur_game, 120, 300, "    ON");
+	if (cur_game->screen.scanlines_on == SDL_TRUE) draw_small_sentence(cur_game, 120, 300, "  >");
+	draw_small_sentence(cur_game, 120, 320, "    OFF");
+	if (cur_game->screen.scanlines_on == SDL_FALSE) draw_small_sentence(cur_game, 120, 320, "  >");
+	/* draw resolution options */
+	draw_small_sentence(cur_game, 420, 160, "RESOLUTIONS");
+	draw_small_sentence(cur_game, 420, 180, "    DEFAULT");
+	if (cur_game->screen.displaymode == 0) draw_small_sentence(cur_game, 420, 180, "  >");
+	draw_small_sentence(cur_game, 420, 200, "    1280x720");
+	if (cur_game->screen.w == 1280 && cur_game->screen.h == 720) draw_small_sentence(cur_game, 420, 200, "  >");
+	draw_small_sentence(cur_game, 420, 220, "    1360x768");
+	if (cur_game->screen.w == 1360 && cur_game->screen.h == 768) draw_small_sentence(cur_game, 420, 220, "  >");
+	draw_small_sentence(cur_game, 420, 240, "    1366x768");
+	if (cur_game->screen.w == 1366 && cur_game->screen.h == 768) draw_small_sentence(cur_game, 420, 240, "  >");
+	draw_small_sentence(cur_game, 420, 260, "    1440x900");
+	if (cur_game->screen.w == 1440 && cur_game->screen.h == 900) draw_small_sentence(cur_game, 420, 260, "  >");
+	draw_small_sentence(cur_game, 420, 280, "    1600x900");
+	if (cur_game->screen.w == 1600 && cur_game->screen.h == 900) draw_small_sentence(cur_game, 420, 280, "  >");
+	draw_small_sentence(cur_game, 420, 300, "    1680x1050");
+	if (cur_game->screen.w == 1680 && cur_game->screen.h == 1050) draw_small_sentence(cur_game, 420, 300, "  >");
+	draw_small_sentence(cur_game, 420, 320, "    1920x1080");
+	if (cur_game->screen.w == 1920 && cur_game->screen.h == 1080) draw_small_sentence(cur_game, 420, 320, "  >");
+	draw_small_sentence(cur_game, 420, 340, "    1920x1200");
+	if (cur_game->screen.w == 1920 && cur_game->screen.h == 1200) draw_small_sentence(cur_game, 420, 340, "  >");
+	draw_small_sentence(cur_game, 420, 360, "    2560x1440");
+	if (cur_game->screen.w == 2560 && cur_game->screen.h == 1440) draw_small_sentence(cur_game, 420, 360, "  >");
+	draw_small_sentence(cur_game, 420, 380, "    2560x1600");
+	if (cur_game->screen.w == 2560 && cur_game->screen.h == 1600) draw_small_sentence(cur_game, 420, 380, "  >");
+	draw_small_sentence(cur_game, 420, 400, "    3840x2160");
+	if (cur_game->screen.w == 3840 && cur_game->screen.h == 2160) draw_small_sentence(cur_game, 420, 400, "  >");
 }
 
 static void
@@ -659,8 +688,25 @@ draw_yesno_screen(struct game *cur_game, char *message)
 	/* Draw background */
 	draw_bg(cur_game);	
 	/* Draw message */
-	draw_sentence(cur_game, 100 * cur_game->screen.scale_x, 100 * cur_game->screen.scale_y, message);
+	draw_sentence(cur_game, 100, 100, message);
 	/* Draw options */
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 150 * cur_game->screen.scale_y, "YES");
-	draw_small_sentence(cur_game, 120 * cur_game->screen.scale_x, 170 * cur_game->screen.scale_y, "NO");
+	draw_small_sentence(cur_game, 120, 150, "YES");
+	draw_small_sentence(cur_game, 120, 170, "NO");
+}
+
+static SDL_bool
+change_resolution(struct game *cur_game, int w, int h)
+{
+	if (cur_game->screen.w == w && cur_game->screen.h == h) {
+		/* Clicking on already-set resolution */
+		return SDL_FALSE;
+	} else {
+		/* If they are in fullscreen desktop mode, change to fullscreen mode */
+		if (cur_game->screen.displaymode == 0) {
+			cur_game->screen.displaymode = 1;
+		}
+		cur_game->screen.w = w;
+		cur_game->screen.h = h;
+		return SDL_TRUE;
+	}
 }