+++ src/loot.c
@@ -490,7 +490,7 @@ swap_item(struct game *cur_game, struct worldmap *map, struct player *cur_player
 					break;
 			}
 		} else if (event.type == SDL_MOUSEBUTTONDOWN) {
-			mouse_click(cur_game, map, cur_player, event.button.x, event.button.y, event.button.button);
+			mouse_click(cur_game, map, cur_player, event.button.x / cur_game->screen.scale_x, event.button.y / cur_game->screen.scale_y, event.button.button);
 			finished = handle_swap(cur_game, cur_player, start_pos);
 		}
 	}